Change balance to Integer

+++ sinatra-obvio/db/migrate/20250520230856_create_accounts.rb
@@ -2,7 +2,7 @@ class CreateAccounts < ActiveRecord::Migration[7.2]
   def change
     create_table :accounts do |t|
       t.integer :id_account    # Columna entera para un id personalizado (opcional)
-      t.float :balance        # Columna para saldo
+      t.integer :balance        # Columna para saldo
       t.string :email         # Columna para email
       t.string :username      # Columna para nombre de usuario
       t.string :password      # Columna para contraseña
